feat(db): implement seeding functionality for various entities
- Add seeding scripts for friends, games, groups, locations, matches, players, scoresheets, sharing, and users to populate the database with test data.
- Introduce new files for each seeding script to modularize the seeding process and improve maintainability.
- Ensure that the seeding process resets relevant tables before inserting new data, maintaining data integrity during tests.
