+++ packages/db/package.json
@@ -46,6 +46,7 @@
     "@board-games/eslint-config": "workspace:*",
     "@board-games/prettier-config": "workspace:*",
     "@board-games/tsconfig": "workspace:*",
+    "@clerk/backend": "^1.34.0",
     "@faker-js/faker": "^9.6.0",
     "@types/d3": "^7.4.3",
     "d3": "^7.9.0",
